requirements.txt
django~=4.0.2
djangorestframework
django-cors-headers
requests~=2.27.1
djangorestframework-simplejwt~=5.1.0